implement `latestSupportedSettingsVersion`
this allows us to parse the latest supported version of settings
that we handle, even if user has newer versions available
(that our current version does not support yet,
 because e.g. didn't get updated yet, etc)

this allows to e.g. develop w/ a new version of the settings,
while e.g. in production, it uses the old settings still.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>